tests/submission/Infraware/Session History/contents/Session_History/Not Use/history_go_zero.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Session History</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>history_go_zero</h3>
		<li>Click 'go(0)' button.</li>
	</p>
	
	<hr>

	<div id="log"></div>

	<iframe id="frame" style="width: 0px; height: 0px;"></iframe>
	<button onclick='onClickBtn()'>go(0)</button>
	
	<script>
		var t1 = async_test('history go zero test');
		
		iframeWindow = document.getElementById('frame').contentWindow;
		var framestr = iframeWindow.localStorage.getItem("load");
				
		function onClickBtn() {
			iframeWindow.localStorage.setItem("load", "true");
			window.history.go(0);
		}
		
		function onLoadEvent(e) {
			if (framestr != "true") return;
			
			t1.step(function() {
				assert_true(true, "load event is raised.");
			});
			t1.done();
			
			iframeWindow.localStorage.setItem("load", "false");
		}

		if (window.addEventListener) {
            window.addEventListener('load', onLoadEvent, false);
        }
        else if (window.attachEvent) {
            window.attachEvent("onload", onLoadEvent);
        }
        else {
			test(function() {
				assert_unreached("DOM event subscribtion is not supported.");
			});
        }
	</script>
 </body>
</html>
